+++ common/samcard.h
@@ -16,21 +16,11 @@
 // Utility functions used in many places, not specific to any piece of code.
 //-----------------------------------------------------------------------------
 
-#ifndef __CARDHELPER_H
-#define __CARDHELPER_H
+#ifndef __SAMCARD_H
+#define __SAMCARD_H
 
 #include <ctype.h>
 #include "common.h"
 
 bool IsHIDSamPresent(bool verbose);
-bool IsCardHelperPresent(bool verbose);
-bool Encrypt(uint8_t *src, uint8_t *dest);
-bool Decrypt(uint8_t *src, uint8_t *dest);
-void DecodeBlock6(uint8_t *src);
-uint8_t GetNumberBlocksForUserId(uint8_t *src);
-uint8_t GetPinSize(uint8_t *src);
-
-int GetConfigCardByIdx(uint8_t typ, uint8_t *blocks);
-int GetConfigCardStrByIdx(uint8_t typ, uint8_t *out);
-int VerifyRdv4Signature(uint8_t *memid, uint8_t *signature);
 #endif
